Now catching any exceptions that occur on recv
Under Windows, sometimes an exception is thrown when using recv() --
this seems to propagate through all the networking libraries. It appears
to have something to do with closing the socket.

Right now, we will be catching all these exceptions and just act as if
the socket is closed in such an occasion.